requirements.txt
django-bootstrap5
Pillow